Move stats ui classes to their own package.
